requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.6
sqlalchemy==1.4.49
jupyter==1.0.0
alembic==1.11.1